Reorganize main window scripts for modularity
